+++ src/test/php/net/daringfireball/markdown/unittest/ReaderInputTest.class.php
@@ -1,0 +1,19 @@
+<?php namespace net\daringfireball\markdown\unittest;
+
+/**
+ * Tests Reader input
+ */
+class ReaderInputTest extends InputTest {
+
+  /**
+   * Create and return new fixture.
+   *
+   * @param   string source
+   * @return  net.daringfireball.markdown.Input
+   */
+  protected function newFixture($source) {
+    return new \net\daringfireball\markdown\ReaderInput(
+      new \io\streams\TextReader(new \io\streams\MemoryInputStream($source))
+    );
+  }
+}